+++ src/containers/QnaPage/index.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 const QnaPage: React.FC = () => {
     return (
         <MainLayout>
-            <h1>Q&A 페이지입니다.</h1>
+            <h1>준비중입니다.</h1>
         </MainLayout>
     );
 };
