--- src/containers/index.tsx
@@ -1,5 +1,0 @@
-export { default as MainPage } from './MainPage';
-export { default as RankPage } from './RankPage';
-export { default as UserPage } from './UserPage';
-export { default as NotFoundPage } from './NotFoundPage';
-export { default as LoginPage } from './LoginPage';
